Repo update

+++ app/src/main/java/com/weatherapp/domain/model/CurrentWeather.kt
@@ -1,8 +1,8 @@
 package com.weatherapp.domain.model
 
 data class CurrentWeather(
-    var temperature: Double? = null,
-    var maximumTemp: Double? = null,
-    var minimumTemp: Double? = null,
+    var temperature: Int? = null,
+    var maximumTemp: Int? = null,
+    var minimumTemp: Int? = null,
     var description: String? = null
 )
