+++ app/src/main/java/com/weatherapp/domain/model/ForecastWeather.kt
@@ -3,7 +3,7 @@ package com.weatherapp.domain.model
 import java.util.*
 
 data class ForecastWeather(
-    var temperature: Double? = null,
+    var temperature: Int? = null,
     var date: Date? = null,
     var description: String? = null
 )
